[wtfos] setup the gpio block/pins and usart1 clock for USART1.
The existing clock/gpio code is fine for this, but if I try to pull
in the USART code I end up quickly needing DMA and other bits, and
I am not ready for that just yet.

So I'll go and write my own USART1 polling driver, then an interrupt
driver, and use that to bootstrap the interrupt manager code.