[wtfos] in progress UART code.
I have no idea if it works, but it doesn't hang the gpio blinky.
In theory I'm now writing an 'a' to the USART1 port every time
I push the button, and that polling check in the USART code
would hang if I am not poking at the USART at all right.

Next up, finding a 3.3v fdti widget and seeing if I'm getting
any 'a's.