[wtfos] [stm32] migrate to my own reimplementation of the stm32f429 startup path
This is a kind of work in progress of stuff that should likely eventually
live in the stm32f429_hw_rcc.c code.  In any case, I've figured out
what the heck the startup path code is doing and implemented a hopefully
cleaner version of it here.

It's still hard-coded and I still need to go and put the configuration
variables for the PLL, clock source, external crystal frequency, etc
somewhere sensible.  But for now it works and it's one less thing that
relies on the STM32 HAL code.