[wtfos] [stm32] hey now I can wait and signal processes
This is still all slow because of my lack of timer framework
in wtfos, but at least now I can put a process to sleep until
a signal comes along and then wake it up.

I hate, HATE, how I'm using spinlocks/critical sections everywhere.
Yes I do just want to get this thing running and somewhat stable
before I go and try to make things not actually suck, but this
is ACTUALLY hurting my sense of pride.

Anyway:

* add container/offset of macros
* add double linked list type
* rename the context switch poke so it's in the platform namespace,
  and I can thus call it from the task routine
* flesh out enough of the task code to implement a dumb as heck
  scheduler that'll run tasks, and manage sleep/wake/running states.
  Heck it doesn't even have a runlist/sleep list, and doesn't care
  about priorities yet.

It's terrible but it is working, which is saying a lot.  I'll go
and write up a timer driver first so tasks can schedule timer events,
and then I'll go tidy up the task/scheduler code to not be this
terrible.