[stm32] [wtfos] add platform routines for the systick timer
In prep for writing the kernel timer side of things, let's hide
the systick stuff behind the platform_* namespace.

That way I can call into it from the kernel routines and it
doesn't need to know about the dirty details.

The interrupt setup and hclk config is still done in the board
specific init file.