[kern] add a hook to call into the timer task if we're idle
This is #if 0'ed out right now because I need to improve the scheduler
a TINY bit with a runlist so I know whether to keep running the timer
or not.  Otherwise if I shut down the timer during idle, I never start
it up again after idle determines it wants to run another task.

Whoops.