[stm32] [kern] ok ok, now we have a round robin scheduler and tickless
well, "tickless" is .. not quite right.

* add an active list for tasks that are running/active/ready.
* run the first task on the list, put at the bottom of the list.
  this makes it a round robin, no priority scheduler for now.
* enable the timer if we have entries in the timer list / more than
  one task active.  If both are false then we don't NEED the timer -
  notably a single task running doesn't need to be preempted by the
  timer just to continue running.
* add some gatekeeping around setting up the task scheduler and
  letting it run.

This ... works? strangely enough?  I haven't tested the timer wheel
stuff yet, but the scheduler is happy now sitting in the idle task
and not waking up via the timer, only waking up during a button press,
or when we kick the test task - and yes, it gets scheduled too.