[kern] [arm] start fudging around with how I could potentially start user tasks.
I'm still not sure how to "do" unprivileged tasks and their stacks right now.

BUT the FreeRTOS code here is using PSP for the current stacks, meaning that
the MSP sitting at the top of memory could STILL be used for some things.

So I gotta go think about that a whole bunch more and clean it up before
I continue.