[kern] add support to start a user task
A user task is a task with both an unpriv and a priv stack, and
will set the initial control register value appropriately when
starting.

I'm hoping (!) that when I'm doing dirty dances with SVC kernel calls
that I won't need to do any extra saving/restoring when a userland
thread is in a kernel syscall, as CONTROL for that thread will be
temporarily set to privileged.

Anyway, i've written a test user task that just calls while (1) and
sits in a loop; it seems to run fine with interrupts and my test
kernel task also running.

Next up, SVC.  Oh it's a GLORIOUS mess.