[stm32] [arm] update the syscall code to be a little .. better?
* drop the arg count to r0..r3 for now.  It may be interesting to
  support r4, r5 as syscall arguments and we can always add them
  back later on.
* .. because they were working by magic, not by intent.  The
  syscall code I wrote was pushing arg5, arg6 onto the stack and just
  magically had their values in r4, r5.
* .. BUT those values need to be pushed into the /kernel stack/
  before we call into the C syscall handler.

Note that the arm_m4_syscall_handler routine is getting r0..r3 from
the original exception saved state, and r8 is magically persisting
across that jump.  I .. dislike that a lot and I'm sure (a) it's wrong,
and (b) it's going to bite me.  I reckon I'm going to have to save
the original return address in the task struct or the kernel stack,
not in r8.

So, this works for now.  If I want more than 4 args then yes
I'll have to write some hacky C inline assembly stubs that will
assemble r0..r3 and then r4, r5 as syscall arguments and then
shovel them into the SVC call in the right way.  However, I still
think how r8 is being used needs to be fixed.

(This has been a fun journey figuring out the depths of the SVC
handling hacks needed for ARM!)