[kern] Fix up the timer code to be more robust/defined around add/delete state
These changes are to better define the state of a timer entry, so
we know if it's not on the pending list, if it's on the pending list,
if it's active.  I want to better define this behaviour so we don't
have a repeat of FreeBSD's kernel timer/callout API when it comes to
queue, dequeue, cancellation points, etc.

I think I've captured all the relevant annoyances.

* In theory you can add/delete timers for other tasks, but they
  may fail, esp if we do SMP one day.
* For now do it from the owner task/context, but not the timer callback.
* I'll add a different call to rearm a timer from inside the timer callback
  so things are cleaner.

More info is in the commit.