[kern] Implement a basic physmem allocator
Its dirty and designed for large (say 256 bytes or later) allocations.
Anything smaller should be done by a zone allocator and I may end up
making the minimum allocator here a bit larger.

Use it for allocating the kernel/user stack in our test task.
So far so good.

The code is terrible, incomplete and I should pull it out into a
separate source file so I can super hammer on it using host side
unit tests.

Notably the freelist way of doing this w/ allocation metadata at the
front of the allocation  makes it difficult to do larger aligned
allocations.  I'll finish this up first and test it thoroughly tested
so I can ensure that it's not /broken/, but eventually I will need
to make it much friendlier for aligned allocations - i'll need these
for doing user task loading and to make the MPU support sane.

Anyway.  This is wtfos.  The whole point is WTF.