[kern] [arm] Implement some placeholder user <-> kernel memory copying
This is important for syscalls as, well, we shouldn't REALLY assume
that it's one big shared happy address space.

So add some methods to access it for now, and convert over the
console putsn to use it.  Yeah it is doing it in an inefficient
way and yeah now every byte being written to is doing a lock/unlock,
but it's /more/ correct.  We can make it faster later.