[wtfos] add initial send / complete kernel test tasks
this adds two kernel tasks - one that creates a global named port
to send things to, and one that will attempt to look it up and
send messages to it once a second.

It ... surprisingly works? Like, I'm super surprised here that it
is indeed working.  It's a bit noisy because, well, I'm now actually
using the physmem allocator, but hey ... it /is/ working.

Well, now I guess I can start playing with port, message, task lifecycles
and all of that annoying crap that needs to be done.  Aiee.