[arm] Don't access TCB and other non-task memory in the syscall return path
This came about when bringing up the MPU!

current_task and the TCB are both unavailable once the CPU is put
back into unprivileged mode.  However we didn't notice this before
as the MPU wasn't enforcing it.

Now that I have some uncommitted MPU patches that enforce it, this
showed up!

So instead do the math and setup first, store it in r1, /then/ flip
to unpriv mode - which with the MPU enabled will start enforcing
memory regions again!