[wtfos] Initial hacked up MPU support.
This adds very basic MPU support on the cortex-m4 / STM32 CPU i'm working on.
It's dirty and it's hard-coded but it /does/ now ensure that the userland
tasks are only touching SRAM it should.

* add the basic plumbing for MPU access, even if it's inefficient
* MPU is configured to default to the default hw region in priv mode,
  so I don't need to program a full priviliged mode MPU map!
* add a table per task with the MPU regions
* manually plumb in regions into the table for user tasks
* during context switch disable/switch tasks/enable MPU
* in the task switch C code plumb in the new table!
* 512 byte alignment now required for the user stack!
  (but not for the kerne stack for said task.)

It's all pretty dirty but it works - and as shown in the previous commit,
it ALSO showed that I was accessing kernel memory after I re-enabled
unprivileged mode!

Ok, so now that it works I can start working on the other parts like IPC.
I'll come back to task specific stuff (and libraries, actual program region
layouts in flash, etc) later.