[user] Don't use -mno-pic-data-is-text-relative
With this option set, it puts /all/ references to the rodata section
/through the GOT/.  This is pretty RAM wasteful.

Maybe we'll need this for libraries but we certainly don't need
this for the first PIC binary or two I'll write.

With it off, the only entry in the GOT now is the global count
variable.  The string stuff goes straight to relative offsets
from text -> rodata and everything is OK.

It's not likely a LOT of extra work to support GOT entries that
redirect into the rodata section - and honestly I should support it! -
but hey at least now it won't waste so much RAM when I write some
more test user tasks.

Well, this one is basically ready to be run.  Time to go and
write the kernel user task loader and get it bootstrapped.