[wtfos] Start fleshing out the user exec routines for our XIP path
This is obviously untested, but at least its a start.

* parse the headers, will eventually validate they're not trash/invalid
* setup each of the segments, GOT relocation.

This doesn't do memory allocation and populate the addrs struct; that'll
be the callers responsibility for now?  That way this code doesn't
REALLY care if there's an XIP or non-XIP binary.  (The only differences
would be setting up the rodata segment and text segment as allocated+
copied, rather than just directly pointing addresses at them.)

Anyway, next step is writing some code to glue this all together and
get ready for starting a task with r9 pointing to GOT, and see how
quickly it crashes!

(oh and populating the MPU, as right now I don't think I have
accounted for the other segments i needed here..)