[kern] migrate the task memory setup into task_mem_setup_mpu()
This hides the layout of the task memory fields and how they map
to the MPU entries from kern/core/task.c .

It isn't ENTIRELY hiding the memory fields as I right now am
setting them directly via stuff passed in as function arguements.

I need to rethink that a little bit in order to setup the rest
of the memory regions before the task creation is finished and the
task is ready to run.