[ipc] more ipc hacks, mostly to start with the bare stupid pipe design
This is going to be super stupid, no send blocking/notification, just
one receiver, enough to tinker with bringing the one-way pipe up and use
it to prototype a basic console kernel task using IPC to send messages
to it.

The more complicated blocking-sender, two-way IPC pipes will need
to wait, and I think I want to build them on top of this.