[console] [pipe] WIP for a console task, using an IPC pipe.
The goal here is to test out the pipe IPC design by having console
IO messages go over the IPC pipe.

It's still lossy on the sending side because we don't have a way
to notify potentially interested senders that the pipe can have
data pushed into it.

That'll have to come next.  Stay tuned.