[user] Fix the user exec GOT relocation logic to work right
This was amusingly picked up by my effort to enable the MPU code
for flash loaded tasks!

I was not doing the math correct, so the BSS GOT entry was way
outside where it should be.

This fixes it entirely!  Now the rest of the MPU enabling stuff
(at least the first pass) for flash loaded tasks can happen.