[ipc] flesh out the free paths some more
* add a flag indicating whether this is a named port or not
* remove ports from the named list if they're on it
* remove the cyclic references and port lists as appropriate during
  close()

It's in close instead of shutdown in case I want to make it so
a shutdown port can be re-opened again for some reason, but honestly
I may just define shutdown as "it's not going to be freed, but
remove the links." That way things can call shutdown as part of
disconnecting / deregistering its service, wait for port IO to
complete, and then call close().

Still untested, as I don't yet have the connect / disconnect
methods wired up.  That's going to be next.